new layout scheme for application for organization purposes
