+++ renders/teams/robot-info.php
@@ -1,9 +1,11 @@
 <?php
-require_once("config.php");
+require_once("../../config.php");
 require_once(ROOT_DIR . "/classes/tables/core/Teams.php");
 require_once(ROOT_DIR . "/classes/tables/core/Events.php");
 require_once(ROOT_DIR . "/classes/tables/core/Years.php");
-require_once(ROOT_DIR . "/classes/tables/local/ScoutCardInfoKeys.php");
+require_once(ROOT_DIR . "/classes/tables/local/RobotInfo.php");
+require_once(ROOT_DIR . "/classes/tables/local/RobotInfoKeys.php");
+require_once(ROOT_DIR . "/classes/tables/core/Matches.php");
 
 
 $eventId = $_GET['eventId'];
@@ -16,11 +18,8 @@ $event = Events::withId($eventId);
 <!doctype html>
 <html lang="en">
 <head>
-    <title><?php echo $team->Id . ' - ' . $team->Name ?> - Stats</title>
+    <title><?php echo $team->Id . ' - ' . $team->Name ?> - Pits</title>
     <?php require_once(INCLUDES_DIR . 'meta.php') ?>
-    <script src="<?php echo JS_URL ?>Chart.min.js"></script>
-    <link rel="stylesheet" href="<?php echo CSS_URL ?>Chart.min.css">
-    <script src="<?php echo JS_URL ?>chartjs-plugin-annotation.min.js"></script>
 </head>
 <body class="mdl-demo mdl-color--grey-100 mdl-color-text--grey-700 mdl-base">
 <div class="mdl-layout mdl-js-layout mdl-layout--fixed-header">
@@ -28,28 +27,28 @@ $event = Events::withId($eventId);
     $navBarArray = new NavBarArray();
 
     $navBarLinksArray = new NavBarLinkArray();
-    $navBarLinksArray[] = new NavBarLink('Teams', 'team-list.php?eventId=' . $event->BlueAllianceId);
+    $navBarLinksArray[] = new NavBarLink('Teams', TEAMS_URL . 'list.php?eventId=' . $event->BlueAllianceId);
     $navBarLinksArray[] = new NavBarLink('Team ' . $teamId, '', true);
 
     $navBarArray[] = new NavBar($navBarLinksArray);
 
     $navBarLinksArray = new NavBarLinkArray();
-    $navBarLinksArray[] = new NavBarLink('Matches', 'team-matches.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id);
-    $navBarLinksArray[] = new NavBarLink('Robot Info', 'team-robot-info.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id);
-    $navBarLinksArray[] = new NavBarLink('Photos', 'team-photos.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id);
-    $navBarLinksArray[] = new NavBarLink('Stats', 'team-stats.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id, true);
+    $navBarLinksArray[] = new NavBarLink('Matches', TEAMS_URL . 'matches.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id);
+    $navBarLinksArray[] = new NavBarLink('Robot Info', TEAMS_URL . 'robot-info.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id, true);
+    $navBarLinksArray[] = new NavBarLink('Photos', TEAMS_URL . 'photos.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id);
+    $navBarLinksArray[] = new NavBarLink('Stats', TEAMS_URL . 'stats.php?eventId=' . $event->BlueAllianceId . '&teamId=' . $team->Id);
 
     $navBarArray[] = new NavBar($navBarLinksArray);
 
     $additionContent = '';
 
-    $profileMedia = $team->getProfileImage(Years::withId($event->YearId));
+    $robotMedia = $team->getProfileImage(Years::withId($event->YearId));
 
-    if (!empty($profileMedia->FileURI))
+    if (!empty($robotMedia->FileURI))
     {
         $additionContent .=
             '<div style="height: unset" class="mdl-layout--large-screen-only mdl-layout__header-row">
-                  <div class="circle-image" style="background-image: url(' . ROBOT_MEDIA_THUMBS_URL . $profileMedia->FileURI . ')">
+                  <div class="circle-image" style="background-image: url(' . ROBOT_MEDIA_THUMBS_URL . $robotMedia->FileURI . ')">
 
                   </div>
                 </div>';
@@ -119,7 +118,7 @@ $event = Events::withId($eventId);
     $additionContent .=
         '</div>';
 
-    $header = new Header($event->Name, $additionContent, $navBarArray, $event, null, 'admin.php?yearId=' . $event->YearId);
+    $header = new Header($event->Name, $additionContent, $navBarArray, $event, null, ADMIN_URL . 'list.php?yearId=' . $event->YearId);
 
     echo $header->toHtml();
 
@@ -130,55 +129,29 @@ $event = Events::withId($eventId);
 
     <main class="mdl-layout__content">
 
-        <div class="content-grid mdl-grid">
+        <?php
 
-            <?php
+            $array = RobotInfo::getObjects(null, null, $event, $team);
 
-            $keys = ScoutCardInfoKeys::getKeys(null, $event);
-
-            $keyStates = array();
-
-            foreach ($keys as $key)
-                if($key->IncludeInStats)
-                {
-                    $keyStr = str_replace(' ', '', $key->KeyState);
-                    $keyStates[$keyStr] = $key->KeyState;
-                }
-
-            foreach ($keyStates as $keyState => $placeholder)
-            {
-                $keyState = str_replace(' ', '', $keyState);
-                ?>
-                <div class="mdl-cell stats-cell">
-                    <p><?php echo $placeholder ?></p>
-                    <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
-                        <input hidden class="mdl-textfield__input" type="text" value="placeholder">
-                        <select class="mdl-textfield__input" id="<?php echo 'change' . $keyState . 'Item' ?>" name="<?php echo 'change' . $keyState . 'Item' ?>">
-                        </select>
-                        <label class="mdl-textfield__label" for="<?php echo 'change' . $keyState . 'Item' ?>">Item</label>
-                    </div>
-                    <div class="stats-chart">
-                        <canvas id="<?php echo $keyState . 'Chart' ?>"></canvas>
-                    </div>
-                </div>
-                <?php
-            }
-            ?>
-
-            <div class="mdl-cell stats-cell" id="oprDprStats">
-                <h6>OPR / DPR</h6>
-                <div>
-                    <h3>OPR: <span id="opr">0</span></h3><br>
-                    <h3>DPR: <span id="dpr">0</span></h3>
-                </div>
-            </div>
-        </div>
+            $array->toHtml();
+        ?>
 
         <?php require_once(INCLUDES_DIR . 'footer.php') ?>
     </main>
 </div>
 <?php require_once(INCLUDES_DIR . 'bottom-scripts.php') ?>
-<script defer src="<?php echo JS_URL ?>stat-charts.js.php?eventId=<?php echo $event->BlueAllianceId ?>"></script>
-<script defer src="<?php echo JS_URL ?>get-opr.js"></script>
+<?php require_once(INCLUDES_DIR . 'modals.php'); ?>
+<script src="<?php echo JS_URL ?>modify-record.js.php"></script>
+<script>
+    function deleteFailCallback(message)
+    {
+        showToast(message);
+    }
+
+    function deleteSuccessCallback(message)
+    {
+        location.reload();
+    }
+</script>
 </body>
 </html>
